Test Case: Variable length Delimiter
After refactoring the code for this test case, the test case for "EmptyString" fails. This occurs because the string starting with custom regular expression only has been taken care of and therefore the test case for an empty string eventually fails. This can be refactored just by filtering out the Empty String (""). (Done in next Commit)